Manual, change log, and setup updates for 1.0.
Here we go: Fiona 1.0. Fiona is OGR's neater, nicer, no tears API.
It's an attempt to make a vector data package that wouldn't be so
out of place in Python's standard library. One that uses Python
idioms and protocols whenever possible. And has documentation.

Thanks for the help and encouragement, everybody!