Always register all drivers when entering.
Previously, I'd been checking to see if any drivers were registered
and then skipping re-registraion if there were. Problem was that
if a user had imported osgeo.gdal, it would have registered only
the GDAL drivers and Fiona would see them and skip registration of
the OGR drivers it needed.

An utterly dumb bug.

I believe this closes both #80 and #92.