Added a BytesCollection class.
This class is a subclass of Collection whose contructor takes a buffer
that we'd like mapped to a virtual file.  The class creates such a file
and passes its path to the underlying Collection, where it can be read
from.  The class cleans up the virtual file when close is called (and by
contruction when the context manager exits).