Support GDAL 1.x and 2.x from a single sdist
Resolves #359

Achieved by refactoring so that ogrext1 and ogrext2 contain the
only parts of the API that vary between GDAL versions. When making
an sdist, we rename each in turn to `ogrext.pyx` and translate to
a ogrext1.c and ogrext2.c source files. Each of these go in the
sdist.