Addresses #422. (#423)
This issue is related to GDAL's GeoJSON driver.  When attempting to open
a buffer of geojson with a BytesCollection where the vsimem filepath has
no extension, the GeoJSON driver will read the first 6000 bytes of the
buffer and look for certain keys that indicate that the file is GeoJSON,
ESRI JSON or TopoJSON.  If those keys are not present, it assumes the
driver can't open the file.

The test here is against valid GeoJSON where these special keys have
been moved past the first 6000 bytes.  The fix employed is to append a
.json extension to the vsimem filepath in the situation where the user
has specifically requested the GeoJSON driver.  The driver, when this
extension is present, skips the 6000 byte test and assumes it can open it.